refactor dictionary
