+++ src/routes.rs
@@ -1,4 +1,3 @@
-pub mod dictionary;
 pub mod faqs;
 pub mod food;
 pub mod printing;
