+++ types/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod contacts;
 pub mod dictionary;
